requirements.txt
python-telegram-bot>=20,<22
python-dotenv>=1.0.1,<2
gradio_client>=0.15,<1

